GameOSFile: implement main save/load paths, name a bunch of symbols, refine split (GameStats earlier)

+++ src/Black/BaseInfo.cpp
@@ -10,7 +10,7 @@ void GBaseInfo::SetInfoID()
 	GGame::g_game->GameLists.BaseInfos.AddToFirst(this);
 }
 
-GBaseInfo* GBaseInfo::GetInfoPtr(int index)
+GBaseInfo* GBaseInfo::GetInfoPtr(unsigned long index)
 {
 	return GGame::g_game->GameLists.BaseInfos.Get(index);
 }
